complete transaction indexer write
